Refactored logic for extracting multi-org client
