+++ integrations/azure-devops/tests/azure_devops/webhooks/webhook_processors/test_advanced_security_webhook_processor.py
@@ -21,8 +21,8 @@ def advanced_security_processor(
     mock_client: MagicMock, monkeypatch: pytest.MonkeyPatch
 ) -> AdvancedSecurityWebhookProcessor:
     monkeypatch.setattr(
-        "azure_devops.client.azure_devops_client.AzureDevopsClient.create_from_ocean_config",
-        lambda: mock_client,
+        "azure_devops.webhooks.webhook_processors.base_processor.AzureDevOpsBaseWebhookProcessor._get_client_for_webhook",
+        lambda self, payload: mock_client,
     )
     return AdvancedSecurityWebhookProcessor(MagicMock(spec=WebhookEvent))
