+++ integrations/azure-devops/tests/azure_devops/webhooks/webhook_processors/test_folder_webhook_processor.py
@@ -12,8 +12,8 @@ def mock_client(monkeypatch: pytest.MonkeyPatch) -> MagicMock:
     mock_client = MagicMock()
     mock_client.get_commit_changes = AsyncMock()
     monkeypatch.setattr(
-        "azure_devops.client.azure_devops_client.AzureDevopsClient.create_from_ocean_config",
-        lambda: mock_client,
+        "azure_devops.webhooks.webhook_processors.base_processor.AzureDevOpsBaseWebhookProcessor._get_client_for_webhook",
+        lambda self, payload: mock_client,
     )
     return mock_client
