+++ integrations/azure-devops/tests/azure_devops/webhooks/webhook_processors/test_work_item_webhook_processor.py
@@ -20,8 +20,8 @@ def work_item_processor(
         return_value={"id": "project-123", "name": "Test Project"}
     )
     monkeypatch.setattr(
-        "azure_devops.client.azure_devops_client.AzureDevopsClient.create_from_ocean_config",
-        lambda: mock_client,
+        "azure_devops.webhooks.webhook_processors.base_processor.AzureDevOpsBaseWebhookProcessor._get_client_for_webhook",
+        lambda self, payload: mock_client,
     )
     return WorkItemWebhookProcessor(event)
 
@@ -177,8 +177,8 @@ async def test_work_item_handle_event_created(
         return_value={"id": "project-123", "name": "Test Project"}
     )
     monkeypatch.setattr(
-        "azure_devops.client.azure_devops_client.AzureDevopsClient.create_from_ocean_config",
-        lambda: mock_client,
+        "azure_devops.webhooks.webhook_processors.base_processor.AzureDevOpsBaseWebhookProcessor._get_client_for_webhook",
+        lambda self, payload: mock_client,
     )
 
     payload = {
@@ -224,8 +224,8 @@ async def test_work_item_handle_event_updated(
         return_value={"id": "project-456", "name": "Test Project"}
     )
     monkeypatch.setattr(
-        "azure_devops.client.azure_devops_client.AzureDevopsClient.create_from_ocean_config",
-        lambda: mock_client,
+        "azure_devops.webhooks.webhook_processors.base_processor.AzureDevOpsBaseWebhookProcessor._get_client_for_webhook",
+        lambda self, payload: mock_client,
     )
 
     payload = {
@@ -269,8 +269,8 @@ async def test_work_item_handle_event_commented(
         return_value={"id": "project-789", "name": "Test Project"}
     )
     monkeypatch.setattr(
-        "azure_devops.client.azure_devops_client.AzureDevopsClient.create_from_ocean_config",
-        lambda: mock_client,
+        "azure_devops.webhooks.webhook_processors.base_processor.AzureDevOpsBaseWebhookProcessor._get_client_for_webhook",
+        lambda self, payload: mock_client,
     )
 
     payload = {
@@ -302,8 +302,8 @@ async def test_work_item_handle_event_deleted(
         return_value={"id": "project-999", "name": "Test Project"}
     )
     monkeypatch.setattr(
-        "azure_devops.client.azure_devops_client.AzureDevopsClient.create_from_ocean_config",
-        lambda: mock_client,
+        "azure_devops.webhooks.webhook_processors.base_processor.AzureDevOpsBaseWebhookProcessor._get_client_for_webhook",
+        lambda self, payload: mock_client,
     )
     payload = {
         "eventType": WorkItemEvents.WORK_ITEM_DELETED,
@@ -394,8 +394,8 @@ async def test_work_item_handle_event_not_found(
         return_value={"id": "project-404", "name": "Test Project"}
     )
     monkeypatch.setattr(
-        "azure_devops.client.azure_devops_client.AzureDevopsClient.create_from_ocean_config",
-        lambda: mock_client,
+        "azure_devops.webhooks.webhook_processors.base_processor.AzureDevOpsBaseWebhookProcessor._get_client_for_webhook",
+        lambda self, payload: mock_client,
     )
 
     payload = {
@@ -427,8 +427,8 @@ async def test_work_item_handle_event_exception(
         return_value={"id": "project-500", "name": "Test Project"}
     )
     monkeypatch.setattr(
-        "azure_devops.client.azure_devops_client.AzureDevopsClient.create_from_ocean_config",
-        lambda: mock_client,
+        "azure_devops.webhooks.webhook_processors.base_processor.AzureDevOpsBaseWebhookProcessor._get_client_for_webhook",
+        lambda self, payload: mock_client,
     )
 
     payload = {
